+++ priv/templates/bonny.gen/init/config.exs
@@ -5,7 +5,7 @@ config :bonny,
 
   # Function to call to get a K8s.Conn object.
   # The function should return a %K8s.Conn{} struct or a {:ok, %K8s.Conn{}} tuple
-  get_conn: {<%= assigns[:app_name] %>.K8sConn, :get, [config_env()]},
+  get_conn: {<%= assigns[:app_name] %>.K8sConn, :get!, [config_env()]},
 
   # Name must only consist of only lowercase letters and hyphens.
   # Defaults to hyphenated mix app name